requirements.txt
pandas==1.1.3
numpy==1.19.5
scikit-learn==0.24.1
imbalanced-learn==0.7.0
scikit-optimize==0.8.1
joblib==1.0.1
flask==2.0.1
pytest==6.2.5
python-dotenv==0.19.0
requests==2.26.0
